requirements/base.txt
# Core requirements for CNPJ Data Pipeline
requests
beautifulsoup4  # For web scraping (replaces bs4)
python-dotenv
polars
psutil
